requirements.txt
python-docx==0.8.11
requests
httpx[http2]==0.13.3
beautifulsoup4==4.11.2
hijridate==2.3.0
tk
tkcalendar==1.6.1
